+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - added support for activating skills in custom commands
 - correctly using the partial file content when using read_file tool with offset and limit
 - added support for both percentage and token count thresholds for context compaction
+- added Smart context compation type
 
 ## [0.64.0]